test: split name validation into per-component test files under components/__tests__
